requirements.txt
beautifulsoup4==4.12.0
Django==4.0.5
Django==4.2
django_braces==1.15.0
djangorestframework==3.14.0
Requests==2.31.0
schedule==1.2.0
websocket-client==0.56.0